Se refactoriza codigo de invoices para separar client component y server component
