+++ src/interfaces/invoices/IInvoices.ts
@@ -1,0 +1,7 @@
+import { IOrderWithStore } from "../orders/IOrderWithStore"
+import { IStore } from "../stores/IStore"
+
+export interface InvoicesClientProps {
+    initialOrders: IOrderWithStore[]
+    stores: IStore[]
+}
